# docs/api-spec.yaml
openapi: "3.0.3"
info:
  title: "AgentX Platform API"
  description: "AgentX 电商 AI 工作助手 API 接口定义"
  version: "1.0.0"
  contact:
    name: "AgentX Team"
servers:
  - url: "http://localhost:8000"
    description: "本地开发环境"
  - url: "https://api.agentx.example.com"
    description: "生产环境"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: boolean
          example: true
        message:
          type: string
        status_code:
          type: integer
        timestamp:
          type: string
          format: date-time

    # ---- 认证 ----
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 1
          example: "zhangsan"
        password:
          type: string
          minLength: 8
          example: "password123"
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: "JWT access token, 30 分钟过期"
        refresh_token:
          type: string
          description: "刷新令牌, 7 天过期"
        token_type:
          type: string
          example: "bearer"
    RegisterRequest:
      type: object
      required: [username, password, company_name, brand_name, category]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 8
        email:
          type: string
          format: email
        company_name:
          type: string
          maxLength: 100
        brand_name:
          type: string
          maxLength: 100
        category:
          type: string
          maxLength: 50
          description: "行业分类"
    UserInfo:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        company_id:
          type: integer
        company_name:
          type: string
        is_admin:
          type: boolean

    # ---- 聊天 ----
...
    KolProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: "李佳琦Austin"
        platform:
          type: string
          enum: [douyin, xiaohongshu, kuaishou, bilibili, weibo]
        followers:
          type: integer
          description: "粉丝数"
          example: 48500000
        engagement_rate:
          type: number
          description: "互动率 (百分比)"
          example: 3.5
        category:
          type: string
          example: "美妆"
        avg_views:
          type: integer
          description: "平均播放量"
        avg_likes:
          type: integer
          description: "平均点赞数"
        avg_comments:
          type: integer
          description: "平均评论数"
        price_range:
          type: string
          description: "报价区间"
          example: "8000-15000"
        location:
          type: string
          description: "所在地"
        verified:
          type: boolean
        bio:
          type: string
          description: "个人简介"
        avatar_url:
          type: string
        created_at:
          type: string
          format: date-time
    KolSearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          example: "美妆达人"
        platform:
          type: string
          enum: [douyin, xiaohongshu, kuaishou, bilibili, weibo, all]
          default: all
        category:
          type: string
        min_followers:
          type: integer
          default: 0
        max_followers:
          type: integer
        min_engagement_rate:
          type: number
          default: 0
        sort_by:
          type: string
          enum: [followers, engagement_rate, relevance]
          default: relevance
        limit:
          type: integer
          minimum: 1
          maximum: 50
          default: 20
    KolSearchResponse:
      type: object
      properties:
        total:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/KolProfile"
        search_id:
          type: string
          description: "搜索 ID，用于后续追问"
    KolExportRequest:
      type: object
      properties:
        kol_ids:
          type: array
          items:
            type: integer
        format:
          type: string
          enum: [csv, pdf]
          default: csv

    # ---- 审核 ----
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/token:
    post:
      tags: [auth]
      summary: "用户登录"
      description: "使用用户名密码登录，返回 JWT access_token 和 refresh_token"
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: "登录成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "401":
          description: "用户名或密码错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/users/register:
    post:
      tags: [auth]
      summary: "用户注册"
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: "注册成功"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserInfo"
        "400":
          description: "用户名已存在"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "422":
          description: "参数校验失败"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/kol/search:
    post:
      tags: [kol]
      summary: "达人搜索"
      description: "搜索达人，支持多平台、多条件筛选"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KolSearchRequest"
      responses:
        "200":
          description: "搜索结果"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KolSearchResponse"
  /api/kol/{kol_id}:
    get:
      tags: [kol]
      summary: "达人详情"
      parameters:
        - name: kol_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: "达人详情"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KolProfile"
        "404":
          description: "达人不存在"
  /api/kol/export:
    post:
      tags: [kol]
      summary: "导出达人列表"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KolExportRequest"
      responses:
        "200":
          description: "导出文件"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # ============================================================
  # 审核确认
  # ============================================================
...